2/24 작업장 정보 애니메이션 효과및 css수정
